// src/app/store/authentication/actions/authentications.actions.types.ts
export enum AuthActionTypes {
  LOGIN = '[AUTH] Login',
  LOGIN_SUCCESS = '[AUTH] Login Success',
  LOGIN_ERROR = '[AUTH] Login Error',
  LOGOUT = '[AUTH] Logout',
  LOG_BACK_IN = '[AUTH] Log back in',
  SIGNUP = '[AUTH] Sign Up',
  SIGNUP_SUCCESS = '[AUTH] Sign Up Success',
  SIGNUP_ERROR = '[AUTH] Sign Up Error'
}